rys_motors_controller: fixed throttle and rotation being int in 1 place causing value clipping

+++ src/rys_base/motors_controller/src/rys_motors_controller.cpp
@@ -24,8 +24,8 @@ volatile float roll;
 volatile float rollPrevious;
 volatile float rotationX;
 
-volatile int rotation;
-volatile int throttle;
+volatile float rotation;
+volatile float throttle;
 
 Motors motors;
 Controller controller;
